fix: move init client to seperate file
