+++ integrations/sonarqube/utils.py
@@ -2,7 +2,6 @@ from typing import Any
 
 from client import SonarQubeClient
 from integration import SonarQubeComponentProjectSelector
-from port_ocean.context.ocean import ocean
 
 
 def produce_component_params(
@@ -17,13 +16,3 @@ def produce_component_params(
     if client.is_onpremise and selector:
         component_query_params.update(selector.generate_request_params())
     return component_query_params
-
-
-def init_sonar_client() -> SonarQubeClient:
-    return SonarQubeClient(
-        ocean.integration_config.get("sonar_url", "https://sonarcloud.io"),
-        ocean.integration_config["sonar_api_token"],
-        ocean.integration_config.get("sonar_organization_id"),
-        ocean.integration_config.get("app_host"),
-        ocean.integration_config["sonar_is_on_premise"],
-    )
